Updated Fog Density and tweaked the ExperimentScript
The fog has been increased for a denser atmosphere, enhancing the sense of immersion. As the camera is now controlled by waypoints, the previous instructions regarding the use of the trigger and joystick for movement have been removed. The transitions between each waypoint now drive the camera's movement, allowing for a smoother and more controlled experience.