Overhauled ExperimentScript
Created StartCS which moves all the code that use to be in CS+ and CS- to a function. Created InterTrial function which moves our code that use to be in a interTrial interval into one function. This changes allow us to make mass changes fast, code more readable, and more consistent.