Added HD Experiment scene
New scene "HD main" is a realistic environment for testing CS+ and CS-. Implemented new ABA technique to apply different material color to objects in the 3d environment World and revert the colors. New "HD ExperimentScript" only change is that it removes previous ABA technique and replaces it with the one built for this 3d environment. ✈️