Major update
Scene Memory (Temporary Name):
In this scene, a set of cubes initially flash their colors before resetting to their original (default) state. After this animation, the user can interact with the cubes using a VR controller and ray-based interaction to change each cube's color manually to match that of the flash. Much more work to be done but a big step forward.