Major Overhaul
The best code is less code. Wrapped major code into functions. Made it so that cubes are no long taken in individually but as an array. Now use "for loops" to do things like display patterns and rest game objects back to the original color making it scalable instead of fixed. File headers now includes ResponseTime. Added a new function to ChangeColorOnSelect called ResetState which resets the state back to what it originally was so that it is ready for the next trial.