Changes
Made the cube grid 5 by 5 for a total of 25 cubes. Cube headers is now adaptable to the amount cubes ie will add more or less cubes to the header of the file based on the length of cubes array. ButtonPressed function now takes in a bool array so that it is now adaptable. Unity was bitching so I added in some code that cleans up the preview render. Display Pattern now just takes in pattern1 instead of a whole new bool array simplifying the architecture.