Updated Backend
The golden rule of programming is to never mess with code that works and that's exactly what I did. Added in the CheckFocusedObject() to the backend of SXR. Problem was that there where discrepancies between the mainfile and the eyetracker files as in there would be hundreds if not thousands of lines missing from the mainfile that were in the eyetracker. So we now save "GazeHitPointX,GazeHitPointY,GazeHitPointZ,GameObjectInFocus" into the eyetracker file now without parentheses. Will soon be updating other scripts to remove checkfocus from them.