First Commit
Problem with first experiment was that it was too simple. This new fear conditioning takes advantage of VR unique features. Monster runs towards player while a high pitch scream plays. monster dynamically adjust based on players position and height. CS is now a blue and red door that open when the US plays.